pixelpipe: check if `piece' is null to avoid crashes
pixelpipe_get_histogram_backbuf() can be called with `piece' set to
NULL, which can lead to crashes under some conditions (e.g. when the
crop filter is active) as it is directly dereferenced without checking
its validity.  Such a crash is mentionned in issue #357.

This adds a null pointer check before dereferencing `piece', making the
function return early if it is not successful.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>