Lighttable: Show image's folder when import starts
This code cleverly selects the directory to display,
ensuring that the lighttable view moves to the appropriate
directory when the first image is imported.
It is configured to also display the contents of all
subdirectories within the selected one.

The scheme is the following:

- If user imports images in place and View mode is on "Tree":
  - if the user selecter 1 folder in Import:
    - the lighttable displays the contents of that folder.
  - else, the lighttable displays the contents of the folder
          showing in the file explorer in Import.

- In all other cases, the lighttable displays the first
imported image's folder.

When the View mode is on "Tree", a `*` is added at
the end of the path to show subfolders content.