Remove hidden shortcut to lighttable by double click
When double-clicking in certain areas of the darkroom (mainly on the border of the central image),
the view would switch to the lighttable. This behavior is no longer expected or necessary,
as there is now a home button, a menu entry to open the lighttable, and a keyboard shortcut for it.