Take the PR https://github.com/aurelienpierreeng/ansel/pull/379 of jakehl and improve it. The MacOS X dedicated build script 2_build_hb_ansel_custom.sh is updated to take into account the new installation directory (in $PROJECT_DIR/install). This script, as well as the 1_install_hb_dependencies.sh script, aren't to be used by the CI. These are for users wishing build Ansel by hand onto their Mac machine. So, the script 2_build_hb_ansel_custom.sh fetches the version of MacOS X to pass it to the cmake statement (instead of given a hard coded value corresponding to the minimum expected version of MacOS X).
Update also the CI script used to check pull requests so that the build of Ansel for MacOS X
is also executed. The Debug build type is removed for MacOS X because the uint64_t type is an
unsigned long long on MacOS X whereas it is an unsigned long in other platforms and the debug
messages in the code use values of type uint64_t (hence the build for the debug build type fails).